fix gitignore, move ssr files
