+++ .gitignore
@@ -1,22 +1,16 @@
 .DS_Store
-node_modules
-compiler
-ssr
-shared.js
-scratch
-!test/compiler
-!test/ssr
 .nyc_output
-coverage
-coverage.lcov
-test/sourcemaps/samples/*/output.js
-test/sourcemaps/samples/*/output.js.map
-_actual.*
-_actual-v2.*
-_actual-bundle.*
-src/compile/shared.ts
-package-lock.json
-.idea/
-*.iml
-store.umd.js
-yarn-error.log
+node_modules
+/compiler/
+/ssr/
+/shared.js
+/scratch/
+/coverage/
+/coverage.lcov/
+/test/sourcemaps/samples/*/output.js
+/test/sourcemaps/samples/*/output.js.map
+/src/compile/shared.ts
+/package-lock.json
+/store.umd.js
+/yarn-error.log
+_actual*.*
